Refactor: 구조변경
