+++ app/components/user/composite/SettingTab.tsx
@@ -1,19 +1,18 @@
-import ProfileSettings from '@/app/components/user/ui/ProfileSettings';
-import PasswoardBox from '@/app/components/user/ui/PasswordSettings';
-import { useFullModal } from '@/app/(core)/user/model/useFullModalStore';
-
-export default function ProfileTab() {
+import ProfileSettingsContainer from '@/app/components/user/container/ProfileSettingsContainer';
+import { useFullModal } from '@/app/(core)/user/store/useFullModalStore';
+import PasswordSettingsContainer from '@/app/components/user/container/PasswordSettingsContainer';
+export default function SettingTab() {
   const open = useFullModal((state) => state.action.open);
 
   return (
     <div>
       {/* 프로필 */}
-      <ProfileSettings />
+      <ProfileSettingsContainer />
       <hr className="text-gray-4 my-3" />
       <div className="flex flex-col gap-3 text-sm">
         <button
           className="border-gray-5 text-gray-5 hover:text-main hover:border-main cursor-pointer rounded-lg border p-1"
-          onClick={() => open(<PasswoardBox />)}
+          onClick={() => open(<PasswordSettingsContainer />)}
         >
           비밀번호 변경
         </button>
