+++ app/components/user/presentation/PeriodBarChart.tsx
